Restructure metrix packages
