+++ hmrc-mongo-metrix-play-26/src/main/scala/uk/gov/hmrc/mongo/metrix/MetricOrchestrator.scala
@@ -19,9 +19,7 @@ package uk.gov.hmrc.mongo.metrix
 import com.codahale.metrics.MetricRegistry
 import play.api.Logger
 import uk.gov.hmrc.mongo.lock.MongoLockService
-import uk.gov.hmrc.mongo.metrix.cache.MetricCache
-import uk.gov.hmrc.mongo.metrix.domain.{MetricRepository, MetricSource, PersistedMetric}
-import uk.gov.hmrc.mongo.metrix.gauge.CachedMetricGauge
+import uk.gov.hmrc.mongo.metrix.internal.{CachedMetricGauge, MetricCache}
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -50,10 +48,10 @@ final case class MetricsOnlyRefreshed(refreshedMetrics: List[PersistedMetric]) e
 }
 
 class MetricOrchestrator(
-  metricSources: List[MetricSource],
-  lockService: MongoLockService,
+  metricSources   : List[MetricSource],
+  lockService     : MongoLockService,
   metricRepository: MetricRepository,
-  metricRegistry: MetricRegistry
+  metricRegistry  : MetricRegistry
 ) {
 
   val metricCache = new MetricCache()
