+++ hmrc-mongo-metrix-play-26/src/main/scala/uk/gov/hmrc/mongo/metrix/MetricSource.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.mongo.metrix.domain
+package uk.gov.hmrc.mongo.metrix
 
 import scala.concurrent.{ExecutionContext, Future}
