+++ hmrc-mongo-metrix-play-26/src/test/scala/uk/gov/hmrc/mongo/metrix/impl/MongoMetricRepositorySpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.mongo.metrix.persistence
+package uk.gov.hmrc.mongo.metrix.impl
 
 import com.mongodb.BasicDBObject
 import org.mongodb.scala.{MongoClient, MongoDatabase}
@@ -22,7 +22,7 @@ import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{BeforeAndAfterEach, LoneElement}
 import uk.gov.hmrc.mongo.component.MongoComponent
 import uk.gov.hmrc.mongo.metrix.UnitSpec
-import uk.gov.hmrc.mongo.metrix.domain.PersistedMetric
+import uk.gov.hmrc.mongo.metrix.PersistedMetric
 import uk.gov.hmrc.mongo.test.MongoSupport
 
 import scala.concurrent.ExecutionContext.Implicits.global
